use chrono for walltime since it is thread safe
